requirements.txt
azure-monitor-query>=1.2.0
azure-identity>=1.15.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
openai>=1.0.0
rich>=13.0.0
flask>=3.0.0
flask-cors>=4.0.0
